Move week_of into own module
