+++ components/calendar/src/lib.rs
@@ -151,7 +151,6 @@
 extern crate alloc;
 
 pub mod any_calendar;
-pub mod arithmetic;
 pub mod buddhist;
 mod calendar;
 mod calendar_arithmetic;
@@ -168,6 +167,7 @@ pub mod japanese;
 pub mod julian;
 pub mod provider;
 pub mod types;
+pub mod week_of;
 
 pub use any_calendar::{AnyCalendar, AnyCalendarKind};
 pub use calendar::Calendar;
