+++ components/datetime/src/input.rs
@@ -8,7 +8,7 @@
 use crate::provider::time_zones::{MetaZoneId, TimeZoneBcp47Id};
 use icu_calendar::any_calendar::AnyCalendarKind;
 use icu_calendar::Calendar;
-use icu_calendar::{arithmetic::week_of, AsCalendar, Date, DateTime, Iso};
+use icu_calendar::{week_of, AsCalendar, Date, DateTime, Iso};
 use icu_timezone::{CustomTimeZone, GmtOffset, TimeVariant};
 
 // TODO (Manishearth) fix up imports to directly import from icu_calendar
